renaming feature 2 to completion_time_analysis
