+++ completion_time_analysis.py
